requirements.txt
langchain==0.3.7 
langchain-ollama 
langchain-chroma 
langchain-community 
flashrank 
python-dotenv 
streamlit 
beautifulsoup4
tqdm